PR #20332: [ds-fusion] Add runtime support for host calculation of offsets in ds fusion
Imported from GitHub PR https://github.com/openxla/xla/pull/20332

This patch adds the support for calculating offset on the host at runtime when the offset depends on the loop induction variable. This is done by extracting the offset computation, the induction variable initialization and the induction variable update as independent computations and they are evaluated on the host at runtime. This avoids device-to-host copy for this fusion in these cases.
Copybara import of the project:

--
1b511365f20fcff612b0a69bfe09572d286f247f by Shraiysh Vaishay <svaishay@nvidia.com>:

Add runtime support for host calculation of offsets in ds fusion

This patch adds the support for calculating offset on the host at
runtime when the offset depends on the loop induction variable. This is
done by extracting the offset computation, the induction variable
initialization and the induction variable update as independent
computations and they are evaluated on the host at runtime. This avoids
device-to-host copy for this fusion in these cases.

--
28d6b62f84bfa14c36e5e49942f5248724715f98 by Shraiysh Vaishay <svaishay@nvidia.com>:

Addressed comments

--
956574290958ce78d2d91ffc518d605a5669dc97 by Shraiysh Vaishay <svaishay@nvidia.com>:

Rebase

--
9a0b89b3382e186b5b15d8432df4037904806939 by Shraiysh Vaishay <svaishay@nvidia.com>:

Rebase

Merging this change closes #20332

FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/openxla/xla/pull/20332 from shraiysh:ds_fusion_3 9a0b89b3382e186b5b15d8432df4037904806939
PiperOrigin-RevId: 725220980