[XLA:CPU] Use Intrinsic::Type everywhere to reduce boilerplate. Fix simd names.
Simd names were colliding for functions with the same vector width but differing types, e.g.  rsqrt(<4 x double) vs rsqrt(<4 x float). In fixing that, I unified name generation and switched to Eugene's new Intrinsic::Type.
Function and simd names are now generated uniformly based on FunctionName() and a span of Intrinsic::Types passed in.
SupportedVectorTypes returns a vector of vectors of types, possible function invocations. The nesting is helpful for functions like ldexp that take two arguments, though we could also have just mandated that all arguments be deducible from the first (which in ldexp is true: the 2nd arg is always i32 and the same width as the first arg).

PiperOrigin-RevId: 783049644